Add validation and exception handling

+++ src/main/java/com/bookingbustickets/bookingbustickets/domain/service/PassengerCategoryService.java
@@ -3,7 +3,7 @@ package com.bookingbustickets.bookingbustickets.domain.service;
 import com.bookingbustickets.bookingbustickets.controller.request.RequestPassengerCategoryDto;
 import com.bookingbustickets.bookingbustickets.domain.model.PassengerCategory;
 import com.bookingbustickets.bookingbustickets.domain.repository.PassengerCategoryRepository;
-import org.springframework.dao.EmptyResultDataAccessException;
+import com.bookingbustickets.bookingbustickets.exception.PassengerCategoryNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -23,7 +23,7 @@ public class PassengerCategoryService {
     public PassengerCategory findPassengerCategoryById(Long id) {
         Optional<PassengerCategory> optionalPassengerCategory = passengerCategoryRepository.findById(id);
         if (optionalPassengerCategory.isEmpty()) {
-            throw new RuntimeException("Passenger Category with ID " + id + " does not exist");
+            throw new PassengerCategoryNotFoundException("Passenger Category with ID " + id + " is not found");
         }
         return optionalPassengerCategory.get();
     }
@@ -41,10 +41,10 @@ public class PassengerCategoryService {
     }
 
     public void deletePassengerCategory(Long id) {
-        try {
+        if (passengerCategoryRepository.existsById(id)) {
             passengerCategoryRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new RuntimeException("Passenger Category with id " + id + "does not exist");
+        } else {
+            throw new PassengerCategoryNotFoundException("Passenger Category with ID " + id + " is not found");
         }
     }
