+++ src/main/java/com/bookingbustickets/bookingbustickets/domain/service/ReservationService.java
@@ -2,7 +2,7 @@ package com.bookingbustickets.bookingbustickets.domain.service;
 
 import com.bookingbustickets.bookingbustickets.domain.model.Reservation;
 import com.bookingbustickets.bookingbustickets.domain.repository.ReservationRepository;
-import org.springframework.dao.EmptyResultDataAccessException;
+import com.bookingbustickets.bookingbustickets.exception.ReservationNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -24,7 +24,7 @@ public class ReservationService {
     public Reservation findReservationById(Long id) {
         Optional<Reservation> optionalReservation = reservationRepository.findById(id);
         if (optionalReservation.isEmpty()) {
-            throw new RuntimeException("Reservation with ID " + id + " does not exist");
+            throw new ReservationNotFoundException("Reservation with ID " + id + " is not found");
         }
         return optionalReservation.get();
     }
@@ -47,10 +47,10 @@ public class ReservationService {
     }
 
     public void deleteReservation(Long id) {
-        try {
+        if (reservationRepository.existsById(id)) {
             reservationRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new RuntimeException("Reservation with id " + id + "does not exist");
+        } else {
+            throw new ReservationNotFoundException("Reservation with ID " + id + " is not found");
         }
     }
