+++ src/main/java/com/bookingbustickets/bookingbustickets/domain/service/ScheduleService.java
@@ -5,8 +5,8 @@ import com.bookingbustickets.bookingbustickets.domain.model.Route;
 import com.bookingbustickets.bookingbustickets.domain.model.Schedule;
 import com.bookingbustickets.bookingbustickets.domain.repository.RouteRepository;
 import com.bookingbustickets.bookingbustickets.domain.repository.ScheduleRepository;
+import com.bookingbustickets.bookingbustickets.exception.ScheduleNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -27,7 +27,7 @@ public class ScheduleService {
     public Schedule findScheduleById(Long id) {
         Optional<Schedule> optionalSchedule = scheduleRepository.findById(id);
         if (optionalSchedule.isEmpty()) {
-            throw new RuntimeException("Schedule with ID " + id + " is not found");
+            throw new ScheduleNotFoundException("Schedule with ID " + id + " is not found");
         }
         return optionalSchedule.get();
     }
@@ -37,7 +37,7 @@ public class ScheduleService {
         if (optionalRoute.isEmpty()) {
             throw new RuntimeException("Route with the given ID is not found");
         }
-        Schedule createdSchedule = new Schedule(requestScheduleDto.getScheduleDate(), requestScheduleDto.getDepartureTime(), requestScheduleDto.getArrivalTime(),optionalRoute.get());
+        Schedule createdSchedule = new Schedule(requestScheduleDto.getScheduleDate(), requestScheduleDto.getDepartureTime(), requestScheduleDto.getArrivalTime(), optionalRoute.get());
         return scheduleRepository.save(createdSchedule);
     }
 
@@ -50,10 +50,10 @@ public class ScheduleService {
     }
 
     public void deleteSchedule(Long id) {
-        try {
+        if (scheduleRepository.existsById(id)) {
             scheduleRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new RuntimeException("Schedule with ID " + id + "does not exist");
+        } else {
+            throw new ScheduleNotFoundException("Schedule with ID " + id + " is not found");
         }
     }
 }
