+++ src/main/java/com/bookingbustickets/bookingbustickets/domain/service/TicketService.java
@@ -3,12 +3,11 @@ package com.bookingbustickets.bookingbustickets.domain.service;
 import com.bookingbustickets.bookingbustickets.controller.request.RequestTicketDto;
 import com.bookingbustickets.bookingbustickets.domain.model.*;
 import com.bookingbustickets.bookingbustickets.domain.repository.*;
-import org.springframework.dao.EmptyResultDataAccessException;
+import com.bookingbustickets.bookingbustickets.exception.TicketNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.Optional;
 
@@ -36,7 +35,7 @@ public class TicketService {
     public Ticket findTicketById(Long id) {
         Optional<Ticket> optionalTicket = ticketRepository.findById(id);
         if (optionalTicket.isEmpty()) {
-            throw new RuntimeException("Ticket with ID " + id + "does not exist");
+            throw new TicketNotFoundException("Ticket with ID " + id + " is not found");
         }
         return optionalTicket.get();
     }
@@ -62,29 +61,29 @@ public class TicketService {
     public Ticket createTicket(RequestTicketDto requestTicketDto) {
         Optional<Schedule> optionalSchedule = scheduleRepository.findById(requestTicketDto.getScheduleDateId());
         if (optionalSchedule.isEmpty()) {
-            throw new RuntimeException("Schedule with the given ID is not found");
+            throw new TicketNotFoundException("Schedule with the given ID is not found");
         }
 
         Optional<Reservation> optionalReservation = reservationRepository.findById(requestTicketDto.getReservationId());
         if (optionalReservation.isEmpty()) {
-            throw new RuntimeException("Reservation with the given ID is not found");
+            throw new TicketNotFoundException("Reservation with the given ID is not found");
         }
 
         Optional<PassengerCategory> optionalPassengerCategory = passengerCategoryRepository.findById(requestTicketDto.getPassengerCategoryId());
         if (optionalPassengerCategory.isEmpty()) {
-            throw new RuntimeException("Passenger category with the given ID is not found");
+            throw new TicketNotFoundException("Passenger category with the given ID is not found");
         }
 
         Optional<Route> optionalOneWayRoute = routeRepository.findById(requestTicketDto.getOneWayRouteId());
         if (optionalOneWayRoute.isEmpty()) {
-            throw new RuntimeException("One-way route with the given ID is not found");
+            throw new TicketNotFoundException("One-way route with the given ID is not found");
         }
 
         Optional<Route> optionalReturnRoute;
         if (requestTicketDto.getReturnRouteId() != null) {
             optionalReturnRoute = routeRepository.findById(requestTicketDto.getReturnRouteId());
             if (optionalReturnRoute.isEmpty()) {
-                throw new RuntimeException("Return route with the given ID is not found");
+                throw new TicketNotFoundException("Return route with the given ID is not found");
             }
         } else {
             optionalReturnRoute = Optional.empty();
@@ -102,11 +101,11 @@ public class TicketService {
         return ticketRepository.save(ticket);
     }
 
-    public void deleteTicket(@PathVariable("id") Long id) {
-        try {
+    public void deleteTicket(Long id) {
+        if (ticketRepository.existsById(id)) {
             ticketRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new RuntimeException("Ticket with ID " + id + "does not exist");
+        } else {
+            throw new TicketNotFoundException("Ticket with ID " + id + " is not found");
         }
     }
 }
